Extract NavigationFeature from EditorEventManager
